move cameraCanvas logic into custom hook
